refactor(cli): extract org context display logic into shared utility
Address PR review feedback by:
- Creating getOrgContext() utility in src/util/cloud.ts to eliminate
  ~60 lines of duplicated code between share.ts and commands/share.ts
- Removing unnecessary dynamic import of getTeamById that was causing
  overhead and potential silent failures
- Removing redundant "On Organization" display; now only shows
  "Sharing to: Organization" once, right before upload
- Updating tests to mock getOrgContext() function

Fixes dynamic import issue and improves maintainability.
