+++ src/commands/share.ts
@@ -18,7 +18,6 @@ import {
 } from '../share';
 import telemetry from '../telemetry';
 import { loadDefaultConfig } from '../util/config/default';
-import { fetchWithProxy } from '../util/fetch/index';
 import type { Command } from 'commander';
 
 export function notCloudEnabledShareInstructions(): void {
@@ -142,37 +141,6 @@ export function shareCommand(program: Command) {
           }
         }
 
-        // Display org context at top (Git-style) - do this before any other output
-        if (cloudConfig.isEnabled()) {
-          try {
-            const response = await fetchWithProxy(`${cloudConfig.getApiHost()}/api/v1/users/me`, {
-              headers: { Authorization: `Bearer ${cloudConfig.getApiKey()}` },
-            });
-            if (response.ok) {
-              const { organization } = await response.json();
-              const currentTeamId = cloudConfig.getCurrentTeamId(organization.id);
-
-              let orgContext = organization.name;
-
-              if (currentTeamId) {
-                try {
-                  const { getTeamById } = await import('../util/cloud');
-                  const team = await getTeamById(currentTeamId);
-                  if (team.name !== organization.name) {
-                    orgContext = `${organization.name} > ${team.name}`;
-                  }
-                } catch {
-                  // Team lookup failed, use just org name
-                }
-              }
-
-              logger.info(chalk.dim(`On ${chalk.cyan(orgContext)}`));
-            }
-          } catch {
-            // Continue without org info
-          }
-        }
-
         // Now show what we're sharing
         if (isEval && eval_) {
           logger.info(`Sharing eval ${eval_!.id}`);
@@ -226,7 +194,6 @@ export function shareCommand(program: Command) {
               cmdObj.showAuth,
             );
 
-            logger.info('');
             let shouldContinue = false;
             try {
               shouldContinue = await confirm({
